\name{addPercentToCount}

\alias{addPercentToCount}

\title{
Add percent-values to a table.
}

\description{
Internal function. Should not be used.
}

\usage{
addPercentToCount(x, y)
}

\arguments{
  \item{x}{
Table containing absolute values.
}
  \item{y}{
Table containing values in percent.
}
}

\value{
A table containing both absolute values and values in percent.
}