\name{freq}
\alias{freq}

\title{
Calculate frequency.}

\description{
Internal function. Should not be used. 
}

\usage{
freq(x) }

\arguments{
  \item{x}{
  factor for which the frequency information should be obtained.
  }
}

\value{
A dataframe comprising the frequency information.
}

\examples{
1#1
}
