\name{psytabs-package}
\alias{psytabs-package}
\alias{psytabs}
\docType{package}
\title{
Produce well-formatted tables for psychological research
}
\description{
Psytabs produces well-formatted tables in .rtf or .html, which largely conform to "psychological style".
}
\details{
\tabular{ll}{
Package: \tab psytabs\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-10-17\cr
License: \tab GPL 2\cr
}
}
\author{
Johannes Beller, Soeren Kliem

Maintainer: Johannes Beller <johannesbeller@gmail.com>
}

\keyword{ package }