\name{psytabs-package}
\alias{psytabs-package}
\alias{psytabs}
\docType{package}
\title{
Produce well-formatted tables for psychological research
}
\description{
Psytabs produces tables conforming to "psychological style" (i.e. APA style) in the .rtf, .html or .doc format.  
}
\details{
\tabular{ll}{
Package: \tab psytabs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-10-04\cr
License: \tab GPL-2\cr
}
}
\author{
Johannes Beller

Maintainer: Johannes Beller <johannesbeller@gmail.com>
}

\keyword{ package }