\name{saveTable}
\alias{saveTable}
\title{
Save a table.
}
\description{
Saves a table either in rtf-format (recommended) or in HTML-format.}
\usage{
saveTable(x, file, HTML = FALSE)
}

\arguments{
  \item{x}{
data.frame as generated by the different xTable functions in this package (for example by the mokkenTable function).}
  \item{file}{
string; filename the table should be saved as.
}
  \item{HTML}{
logical value that toggles whether the file format should be HTML.}
}
\details{
Currently the text in the .rtf-Format is set to an 8-point size to accommodate the sometimes large tables. 
}
\value{
None
}
\examples{
data(myData)
test <- demographicTable(myData$sex, myData$age_group7, percent=FALSE)
test

#saveTable(test, "test.rtf")
}